Move Status UI Model into it's own file
