+++ feature/dvla/src/main/kotlin/uk/gov/govuk/dvla/ui/model/VehicleSummaryUiModel.kt
@@ -8,23 +8,3 @@ internal data class VehicleSummaryUiModel(
     val motStatus: StatusUiModel,
     val menuItems: List<OverflowMenuItem> = emptyList()
 )
-
-internal sealed interface StatusUiModel {
-    data class StatusRow(
-        val statusRowUi: StatusRowUiModel
-    ) : StatusUiModel
-
-    data class CountdownRow(
-        val countdownBarUi: StatusCountdownUiModel
-    ) : StatusUiModel
-
-    data class InfoRow(
-        val infoRowUi: InfoRowUiModel
-    ) : StatusUiModel
-
-    data class LinkRow(
-        val linkRowUi: LinkRowUiModel
-    ) : StatusUiModel
-
-    data object NoStatus : StatusUiModel
-}
